+++ R/utils.R
@@ -38,3 +38,21 @@ clear_error_handler <- function(retvalue = NA) {
   }
 }
 
+as_r_value <- function(x) {
+  if (inherits(x, "python.builtin.object"))
+    py_to_r(x)
+  else
+    x
+}
+
+yoink <- function(package, symbol) {
+  do.call(":::", list(package, symbol))
+}
+
+defer <- function(expr, envir = parent.frame()) {
+  call <- substitute(
+    evalq(expr, envir = envir),
+    list(expr = substitute(expr), envir = parent.frame())
+  )
+  do.call(base::on.exit, list(substitute(call), add = TRUE), envir = envir)
+}